% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-exports.R
\name{fetch}
\alias{fetch}
\title{Fetch interactions from a File object}
\usage{
fetch(
  file,
  range1 = "",
  range2 = "",
  normalization = "NONE",
  count_type = "int",
  join = FALSE,
  query_type = "UCSC",
  type = "df"
)
}
\arguments{
\item{file}{file from which interactions should be fetched.}

\item{range1}{first set of genomic coordinates of the region to be queried.
Accepted formats are UCSC or BED format.
When not provided, genome-wide interactions will be returned.}

\item{range2}{second set of genomic coordinates of the region to be queried.
When not provided, range2 is assumed to be identical to range1.}

\item{normalization}{name of the normalization factors used to balance interactions.
Specify "NONE" to return raw interactions.}

\item{count_type}{data type used to fetch interactions.
Should be "int" or "float"}

\item{join}{join genomic coordinates onto pixels.
When TRUE, interactions will be returned in bedgraph2 format.
When FALSE, interactions will be returned in COO format.
Ignored when type="dense".}

\item{query_type}{type of the queries provided through range1 and range2 parameters.
Types of query supported: "UCSC", "BED".}

\item{type}{interactions format.
Supported formats: "df", "dense".}
}
\value{
a DataFrame or Matrix object with the interactions for the given query.
}
\description{
Fetch interactions from a File object
}
\examples{
f <- File("interactions.hic", 1000)
fetch(f)  # Fetch genome-wide interactions
fetch(f, "chr1")  # Fetch interactions overlapping a symmetric query (chr1)
fetch(f, "chr1:0-10,000,000", "chr2:10,000,000-20,000,000")  # Fetch interactions overlapping an asymmetric query (chr1:chr2)
fetch(f, normalization="VC")  # Fetch VC-normalized interactions
fetch(f, join=TRUE)  # Fetch interactions with genomic coordinates joined to the count table
fetch(f, "chr1\t0\t10000000", query_type="BED")  # Fetch interactions given a query in BED format
fetch(f, type="dense")  # Fetch interactions in dense format (i.e. as a Matrix)
}
