% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-exports.R
\name{File}
\alias{File}
\title{Open a .hic or .cool file for reading}
\usage{
File(path, resolution = NULL, matrix_type = "observed", matrix_unit = "BP")
}
\arguments{
\item{path}{path to the file to be opened (Cooler URI syntax is supported).}

\item{resolution}{matrix resolution. Required when file is multi-resolution (e.g., .hic or .mcool).}

\item{matrix_type}{type of the matrix to be opened. Should be one of "observed", "oe" or "expected".}

\item{matrix_unit}{unit of the matrix to be opened. Should be one of "BP", "FRAG".}
}
\value{
a file handle.
}
\description{
Open a .hic or .cool file for reading
}
\examples{
\dontrun{
File("interactions.cool")
File("interactions.mcool::/resolutions/100000")
File("interactions.mcool", 100000)
File("interactions.hic", 100000)
}
}
