% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-exports.R
\name{hictkR_open}
\alias{hictkR_open}
\title{Open files in .cool, .mcool, .scool, and .hic format}
\usage{
hictkR_open(path, resolution = NULL, cell = NULL)
}
\arguments{
\item{path}{path to the file to be opened (Cooler URI syntax is supported).}

\item{resolution}{resolution of the file to be opened.
Required when file is in .hic format.}

\item{cell}{name of the cell to be opened.
Ignored when file is not in .scool format.}
}
\value{
a file handle of the appropriate type.
}
\description{
Open files in .cool, .mcool, .scool, and .hic format
}
\examples{
hictkR_open("interactions.cool")
hictkR_open("interactions.mcool")
hictkR_open("interactions.scool")
hictkR_open("interactions.mcool", resolution=10000)
hictkR_open("interactions.hic", resolution=10000)
hictkR_open("interactions.scool", cell="id_0001")
}
