% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-exports.R
\name{is_cooler}
\alias{is_cooler}
\title{Test whether a file is in .cool format}
\usage{
is_cooler(path)
}
\arguments{
\item{path}{path to the file to be tested (Cooler URI syntax is supported).}
}
\description{
Test whether a file is in .cool format
}
