% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-exports.R
\name{is_multires_file}
\alias{is_multires_file}
\title{Test whether a file is in .mcool format}
\usage{
is_multires_file(path)
}
\arguments{
\item{path}{path to the file to be tested.}
}
\description{
Test whether a file is in .mcool format
}
