% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-exports.R
\name{is_scool_file}
\alias{is_scool_file}
\title{Test whether a file is in .cool format}
\usage{
is_scool_file(path)
}
\arguments{
\item{path}{path to the file to be tested.}
}
\description{
Test whether a file is in .cool format
}
