% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-exports.R
\name{MultiResFile}
\alias{MultiResFile}
\title{Opena a .mcool file for reading}
\usage{
MultiResFile(path)
}
\arguments{
\item{path}{path to the file to be opened.}
}
\value{
a file handle.
}
\description{
Opena a .mcool file for reading
}
\examples{
\dontrun{
MultiResolutionFile("interactions.mcool")
}
}
