% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-exports.R
\name{SingleCellFile}
\alias{SingleCellFile}
\title{Opena a .scool file for reading}
\usage{
SingleCellFile(path)
}
\arguments{
\item{path}{path to the file to be opened.}
}
\value{
a file handle.
}
\description{
Opena a .scool file for reading
}
\examples{
\dontrun{
SingleCellFile("interactions.scool")
}
}
