% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hictkR-pkg.R
\docType{package}
\name{hictkR-package}
\alias{hictkR}
\alias{hictkR-package}
\title{hictkR: R bindings for hictk}
\description{
This package provides R binding for hictk: a blazing fast toolkit to work with .hic and .cool files
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paulsengroup/hictkR}
  \item \url{https://paulsengroup.github.io/hictkR}
  \item Report bugs at \url{https://github.com/paulsengroup/hictkR/issues}
}

}
\author{
\strong{Maintainer}: Roberto Rossini \email{roberros@uio.no} (\href{https://orcid.org/0000-0003-3096-1470}{ORCID})

}
